Implement FeaturedComment container
